feat: 페이지 타입 분리
