+++ src/pages/follower/_types/index.d.ts
@@ -1,0 +1,17 @@
+interface Follower {
+  userName: string;
+  imageURL: string;
+}
+
+interface UserProfileData {
+  userName: string;
+  description: string;
+  imageURL: string;
+  follower: number;
+  following: number;
+}
+
+export interface FollowerPageProps {
+  userProfileData: UserProfileData;
+  initialFollowerData: Follower[];
+}
